% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_prepare_data.R
\name{load_and_prepare_data}
\alias{load_and_prepare_data}
\title{Load and Prepare Data}
\usage{
load_and_prepare_data(file)
}
\arguments{
\item{file}{Path to the Excel file.}
}
\value{
A data frame of prepared review data.
}
\description{
This function loads review data from an Excel file and prepares it for analysis.
}
