% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ratings.R
\name{summarize_ratings}
\alias{summarize_ratings}
\title{Summarize Ratings}
\usage{
summarize_ratings(data)
}
\arguments{
\item{data}{A data frame with columns 'Airline_Name' and 'Overall_Rating'.}
}
\value{
A data frame with 'Airline_Name' and 'avg_rating'.
}
\description{
This function calculates the average overall rating for each airline.
}
\examples{
data <- data.frame(Airline_Name = c("Airline A", "Airline B"),
                   Overall_Rating = c(5, 3))
summarize_ratings(data)
}
