% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_top_airlines.R
\name{summarize_top_airlines}
\alias{summarize_top_airlines}
\title{Summarize Top Airlines by Rating}
\usage{
summarize_top_airlines(data)
}
\arguments{
\item{data}{A data frame with columns 'Airline_Name' and 'Overall_Rating'.}
}
\value{
A tibble with the top 10 airlines by average rating.
}
\description{
This function summarizes the top airlines based on average overall ratings.
}
