% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_trends.R
\name{visualize_trends}
\alias{visualize_trends}
\title{Visualize Review Trends}
\usage{
visualize_trends(data)
}
\arguments{
\item{data}{A data frame with 'Review_Date' column.}
}
\value{
A ggplot object showing the review trends.
}
\description{
This function visualizes trends in the number of reviews over time.
}
