% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{Analyze Sentiment}
\usage{
analyze_sentiment(reviews, method = "syuzhet")
}
\arguments{
\item{reviews}{A character vector of reviews.}

\item{method}{The sentiment analysis method (default: "syuzhet").}
}
\value{
A data frame with sentiment scores.
}
\description{
This function analyzes sentiment for a given set of reviews.
}
