% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sentiment.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{Analyze Sentiment}
\usage{
analyze_sentiment(text)
}
\arguments{
\item{text}{A character vector of reviews.}
}
\value{
A data frame with sentiment scores.
}
\description{
Computes sentiment scores for a vector of text reviews.
}
\examples{
analyze_sentiment(c("Amazing flight!", "Terrible service."))
}
