% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_keywords.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract Keywords}
\usage{
extract_keywords(data)
}
\arguments{
\item{data}{A data frame with a 'Review' column.}
}
\value{
A data frame with keywords and their frequencies.
}
\description{
This function extracts keywords from review text.
}
