% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_prepare_data.R
\name{load_and_prepare_data}
\alias{load_and_prepare_data}
\title{Load and Prepare Data}
\usage{
load_and_prepare_data(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the CSV file.}
}
\value{
A data frame with prepared data.
}
\description{
This function loads the airline review dataset from a CSV file and prepares it for analysis.
It assumes the \code{Review.Date} column has been pre-cleaned and converts it into a standard \code{Date} format.
It also ensures \code{Overall_Rating} is numeric.
}
\examples{
file_path <- system.file("extdata", "Airline_review.csv", package = "AirlineReviewAnalyzer")
data <- load_and_prepare_data(file_path)
}
