% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_prepare_data.R
\name{load_and_prepare_data}
\alias{load_and_prepare_data}
\title{Load and Prepare Airline Review Data}
\usage{
load_and_prepare_data(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the CSV file.}
}
\value{
A cleaned data frame ready for analysis.
}
\description{
This function loads the airline review data from a specified CSV file
and cleans it for analysis.
}
