% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_trends.R
\name{visualize_trends}
\alias{visualize_trends}
\title{Visualize Trends}
\usage{
visualize_trends(data)
}
\arguments{
\item{data}{A data frame containing columns \code{Review_Date} (date),
\code{Overall_Rating} (numeric), and \code{Airline_Name} (character).}
}
\value{
A ggplot object showing trends in ratings over time by airline.
}
\description{
This function visualizes trends in airline feedback over time by airline.
}
